# swagger.yaml
openapi: 3.0.0
paths:
  /v1:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        "200":
          description: ""
  /v1/memory:
    get:
      operationId: MemoryV1Controller_getAllMemories
      summary: Get All Memories
      parameters: []
      responses:
        "200":
          description: ""
      tags: &a1
        - Memory
    post:
      operationId: MemoryV1Controller_postMemory
      summary: Create Memory
      description: Create a memory manually.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryPostDto"
      responses:
        "201":
          description: The memory was created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResponseDto"
      tags: *a1
  /v1/memory/auto:
    post:
      operationId: MemoryV1Controller_postMemoryAuto
      summary: Auto Create Memory
      description: Uses AI to generate a memory from a url.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoMemoryPostDto"
      responses:
        "201":
          description: The memory was created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResponseDto"
      tags: *a1
  /v1/memory/{uuid}:
    delete:
      operationId: MemoryV1Controller_deleteMemory
      summary: Delete Memory
      parameters:
        - name: uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags: *a1
info:
  title: Raven
  description: ""
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    AutoMemoryPostDto:
      type: object
      properties:
        url:
          type: string
          example: https://pubmed.ncbi.nlm.nih.gov/15114002/
          description: The url to the memory
      required:
        - url
    MemoryResponseDto:
      type: object
      properties:
        uuid:
          type: string
          example: uuid
          description: The unique identifier of the memory.
        url:
          type: string
          example: https://pubmed.ncbi.nlm.nih.gov/15114002/
          description: The url related to the memory
        title:
          type: string
          example: This is a memory
          description: The title of the memory
        summary:
          type: string
          example: This is a summary of the memory
          description: The summary of the memory
        createdAt:
          format: date-time
          type: string
          example: 2021-08-01T00:00:00.000Z
          description: The summary of the memory
      required:
        - uuid
        - title
        - summary
        - createdAt
    MemoryPostDto:
      type: object
      properties:
        url:
          type: string
          example: https://pubmed.ncbi.nlm.nih.gov/15114002/
          description: The url to the memory
        title:
          type: string
          example: The title of the memory
          description: The title of the memory
        summary:
          type: string
          example: The summary of the memory
          description: The summary of the memory
      required:
        - url
        - title
        - summary